+++ design-system/tag-group/examples/2-rounded.tsx
@@ -64,7 +64,6 @@ export default () => (
 				href="http://www.cupcakeipsum.com/"
 				key={sweet}
 				text={sweet}
-				removeButtonLabel="remove"
 			/>
 		))}
 	</TagGroup>