+++ design-system/tag-group/examples/4-remove-button-label.tsx
@@ -6,17 +6,13 @@ import TagGroup from '../src';
 
 export default () => (
 	<div>
-		<TagGroup>
-			<SimpleTag text="Base Tag" testId="standard" />
-			<SimpleTag text="Linked Tag" href="/components/tag" />
-			<SimpleTag text="Rounded Tag" appearance="rounded" />
-			<Tag text="Removable button" />
-		</TagGroup>
-		<TagGroup alignment="end">
+		<TagGroup alignment="start">
 			<SimpleTag text="Base Tag" />
 			<SimpleTag text="Linked Tag" href="/components/tag" />
 			<SimpleTag text="Rounded Tag" appearance="rounded" />
-			<Tag text="Removable button" />
+			<Tag text="Dog tag" removeButtonLabel="Delete" />
+			<Tag text="Cat tag" removeButtonLabel="Delete" />
+			<Tag text="Mouse tag" removeButtonLabel="Delete" />
 		</TagGroup>
 	</div>
 );