+++ linking-platform/smart-card/src/view/FlexibleCard/components/blocks/action-group/action-group-item/index.tsx
@@ -52,6 +52,10 @@ const ActionGroupItem = ({
 		actionProps.icon = undefined;
 	}
 
+	if (item?.entryPointWrapper) {
+		actionProps.wrapper = item.entryPointWrapper;
+	}
+
 	return (
 		<Action
 			asDropDownItem={asDropDownItems}